Remove short-circuit
The docs specify `Div100vh` as a component that can be used as a drop-in
replacement for <div style={{ height: '100vh' }} />.

Current behavior does not match that - if there is no declared `style`
prop, the resizing behavior does not execute and the component does not
work as expected.

This patch allows for the innerHeight fix to run even if no style prop
is present.